UNOMI-250 Refactoring services into seperate Java packages Right now we have a bit of a mess in our services implementation sub-project. All the service implementations are in the package: org.apache.unomi.services.services I propose that we rename this package to: org.apache.unomi.services.impl and that each service gets its own sub-package, for example the RulesServiceImpl would be in: org.apache.unomi.services.impl.rules This would make it much easier to activate/deactivate logging per service implementation, because currently it is tricky to get logging activated properly unless using configuration by class. Having it by package would be a lot better. Utility classes such as ThirdPartyService or the clustering event classes would also be moved with the service sub-package they belong to. In terms of impact this should minimal, only people activating logs should be notified that the class location has changed.
