+++ services/src/main/java/org/apache/unomi/services/impl/events/ThirdPartyServer.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.unomi.services.services;
+package org.apache.unomi.services.impl.events;
 
 import inet.ipaddr.IPAddress;
